+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 11.7
 - HTML add label to Class Search widget in ClassSearchWidget.fnc.php & Widget.php
 - Add SQLLimitForList() function in database.in.php & ListOutput.fnc.php
 - CSS FlatSIS theme: add box-shadow to colorBox in colors.css, stylesheet.css & zresponsive.css
+- Allow display of `$link['add']` (or remove) on PDF or if not allowed to edit in ListOutput.fnc.php
 - Force display of `$link['add']` on PDF or if not allowed to edit in DailyTransactions.php, Percent.php, InputFinalGrades.php & ProgressReports.php
 
 Changes in 11.6